upgrade to d3 v5 to fix/improve connections WIP
