+++ viewer/package.json
@@ -51,6 +51,7 @@
     "chalk": "^2.0.1",
     "copy-webpack-plugin": "4.5.1",
     "css-loader": "^1.0.0",
+    "d3": "^5.7.0",
     "eslint": "^5.1.0",
     "eslint-config-standard": "^12.0.0-alpha.0",
     "eslint-friendly-formatter": "^4.0.1",
